Updated compiler pragmas
pragmas are used to suppress warnings in limited, known scenarios where the code is safe. this commit abstracts the details of warning suppression for any given toolchain to a common set of macros. the macros are in util/compiler.h. changes to other source files are just to include util/compiler.h where necessary and replace the necessary #pragma directives with the relevant macros